+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Upcoming
 Fixed
 =====
 
+- Set command palette input to vi-insert when opened
 - Fix issue with notebook scrolling
 - Fix issue where deleteing the contents of a large cell results in scrolling the notebook out of view
 - Fix ``pageup`` and ``pagedown`` key press detection